Update vJass  Level Up Heroes.txt
Fixed a typo in scope declaration.
'Level' was missing the last 'l' and this was fixed.

Added support template for Max level announcement.
The functions MaxLevel and Announce were added. Announce needs to be fitted with the Display Timed Text function but the implementation prior should be adequate.

Trigger now sleeps until awakened.
This allows the game flow to be controlled after selection. It was required that the game controls player level progression.